src/apps/trusted_host/requirements.txt
annotated-types==0.6.0
anyio==3.7.1
bcrypt==4.0.1
blinker==1.7.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==41.0.5
exceptiongroup==1.1.3
fastapi==0.104.1
Flask==3.0.0
gunicorn==21.2.0
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysql-connector-python==8.2.0
packaging==23.2
paramiko==3.3.1
ping3==4.0.4
protobuf==4.21.12
pycparser==2.21
pydantic==2.5.1
pydantic_core==2.14.3
PyMySQL==1.1.0
PyNaCl==1.5.0
python-dotenv==1.0.0
requests==2.31.0
sniffio==1.3.0
sshtunnel==0.4.0
starlette==0.27.0
typing_extensions==4.8.0
urllib3==1.26.6
Werkzeug==3.0.1
zipp==3.17.0
